Adding my changes to the branch

+++ python-backend/app/api/mines/notification/models/notification.py
@@ -4,6 +4,7 @@ from sqlalchemy.schema import FetchedValue
 from sqlalchemy.dialects.postgresql import UUID
 from app.extensions import db
 from app.api.utils.models_mixins import  Base
+from werkzeug.exceptions import BadRequest, NotFound
 
 
 class Notification( Base):
@@ -12,20 +13,29 @@ class Notification( Base):
     mine_guid = db.Column(UUID(as_uuid=True),db.ForeignKey('mine.mine_guid'))
     idir = db.Column(db.String, nullable=False)
 
+    def __repr__(self):
+        return '<Notification %r>' % self.user_id
+
+    @classmethod
+    def find_subscription(cls, mine_guid):
+        user_name = User().get_user_username()
+        return Notification.query.filter_by(mine_guid=uuid.UUID(mine_guid)).filter_by(
+            idir=user_name).first()
+
     @classmethod
     def create_subscription(cls, mine_guid):
         user_name = User().get_user_username()
-        if Notification.query.filter_by(mine_guid=uuid.UUID(mine_guid)).filter_by(
-            idir=user_name).first() is None:
-            subscription = cls(mine_guid=uuid.UUID(mine_guid), idir=user_name)
-            subscription.save()
-            return subscription
-        return
+        if cls.find_subscription(mine_guid):
+            raise BadRequest('Already subscribed to mine.')
+        subscription = cls(mine_guid=uuid.UUID(mine_guid), idir=user_name)
+        subscription.save()
+        return subscription
 
     @classmethod
     def delete_subscription(cls, mine_guid):
-        user_name = User().get_user_username()
-        subscription_to_delete = Notification.query.filter_by(mine_guid=uuid.UUID(mine_guid)).filter_by(idir=user_name).first()
+        subscription_to_delete = cls.find_subscription(mine_guid)
+        if not subscription_to_delete:
+            raise NotFound("Subscription not found")
         db.session.delete(subscription_to_delete)
         db.session.commit()
         return subscription_to_delete
